+++ src/ORM/Exception/MissingEntityException.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * MissingEntityException file
+ *
+ * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
+ * @link          http://cakephp.org CakePHP(tm) Project
+ * @since         3.0.0
+ * @license       http://www.opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Cake\ORM\Exception;
+
+use Cake\Core\Exception\Exception;
+
+/**
+ * Exception raised when an Entity  could not be found.
+ *
+ */
+class MissingEntityException extends Exception {
+
+	protected $_messageTemplate = 'Entity class %s could not be found.';
+
+}
